* Refactoring the component structure * Update style
