--- src/components/pwc-map-controls/components/control-marker-template/control-marker-template.css
@@ -1,24 +1,0 @@
-.form-group {
-  display: flex;
-}
-.form-group > label {
-  width: 60px;
-  line-height: 30px;
-}
-.form-group > input {
-  flex: 1;
-}
-.form-control {
-  border: none;
-  background: #fff;
-  border-radius: 4px;
-  font-size: 20px;
-}
-
-color-picker {
-  padding-left: 10px;
-}
-
-pwc-ibox-footer {
-  text-align: right;
-}
