--- src/components/pwc-map-controls/components/pwc-map-text-control/pwc-map-text-control.css
@@ -1,35 +1,0 @@
-pwc-ibox {
-  position: absolute;
-  left: 30px;
-}
-.animated {
-  animation-duration: 1s;
-  animation-fill-mode: both;
-}
-
-@-webkit-keyframes fadeInLeft {
-  0% {
-    opacity: 0;
-    -webkit-transform: translateX(-50px);
-  }
-  100% {
-    opacity: 1;
-    -webkit-transform: translateX(0);
-  }
-}
-
-@keyframes fadeInLeft {
-  0% {
-    opacity: 0;
-    transform: translateX(-50px);
-  }
-  100% {
-    opacity: 1;
-    transform: translateX(0);
-  }
-}
-
-.fadeInLeft {
-  -webkit-animation-name: fadeInLeft;
-  animation-name: fadeInLeft;
-}
